funding_assistant/requirements.txt
apscheduler>=3.10.4
google-genai>=1.0.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
requests>=2.31.0
